reset_custom_column_value function

+++ include/global_variables.h
@@ -28,7 +28,7 @@ extern sqlite3_stmt* get_file_columns_statement;
 extern sqlite3_stmt* search_file_from_hash_statement;
 extern sqlite3_stmt* set_filepath_statement;
 
-extern unsigned int custom_columns_n;
+extern unsigned short custom_columns_n;
 extern struct custom_column* custom_columns;
 
 extern char main_path[MAX_PATH];	//path where .nothydrus directory is located
